+++ src/main/java/mykhailok/inventory/service/UserService.java
@@ -3,15 +3,17 @@ package mykhailok.inventory.service;
 import mykhailok.inventory.model.User;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.math.BigInteger;
+
 
 public interface UserService {
 
     void save(User user);
 
     @Transactional
-    void deleteById(int id);
+    void deleteById(BigInteger id);
 
     User findByUsername(String username);
 
-    int findMaxId();
+    BigInteger findMaxId();
 }
